integrated reversioning with instance config

+++ gemsModules/systemoperations/instance_config/versions.py
@@ -8,42 +8,67 @@ import shutil
 class DateReversioner:
     """A class to version a json file based on a timestamp."""
 
-    def __init__(self, active_config: Path, example_config: Path):
+    def __init__(
+        self, active_config: Path, new_config: Path = None, new_config_data: dict = None
+    ):
         self.file_to_version = active_config
-        self.example = example_config
+        self._old_version_data = None
+
+        self._new_version = new_config
+        self._new_version_data = new_config_data
+
+    @property
+    def old_version(self):
+        """Return the old version data."""
+        if self.file_to_version.exists() and self._old_version_data is None:
+            with open(self.file_to_version) as f:
+                self._old_version_data = json.load(f)
+
+        return self._old_version_data
+
+    @property
+    def new_version(self):
+        """Return the new version data."""
+        if self._new_version_data is None and self._new_version is not None:
+            with open(self._new_version) as f:
+                self._new_version_data = json.load(f)
+
+        return self._new_version_data
 
     def update(self):
         """Update the file_to_version with the current date."""
 
         needs_update = False
         if self.file_to_version.exists():
-            # could generalize
-            with open(self.file_to_version) as f:
-                j_data = json.load(f)
-            with open(self.example) as f:
-                tj_data = json.load(f)
+            if "date" not in self.new_version:
+                raise RuntimeError("Template file does not have a date, cannot update.")
 
-            if "date" not in j_data:
+            if "date" not in self.old_version:
                 # if no date is found, lets update to a versioned file
                 needs_update = True
             else:
                 # check the template file for a date
                 # if the template is newer, lets update to a versioned file
-                if "date" in tj_data:
-                    if j_data["date"] < tj_data["date"]:
-                        needs_update = True
-                else:
-                    raise RuntimeError(
-                        "Template file does not have a date, cannot update."
-                    )
+                old_date = datetime.datetime.fromisoformat(self.old_version["date"])
+                new_date = datetime.datetime.fromisoformat(self.new_version["date"])
+                if new_date > old_date:
+                    needs_update = True
+        else:
+            needs_update = True
 
         if needs_update:
             # Backup the file_to_version.
-            backup_file = self.file_to_version.with_suffix(
-                f".{datetime.datetime.now().strftime('%Y%m%d%H%M%S')}.bak"
-            )
-            shutil.copyfile(self.file_to_version, backup_file)
+            if self.file_to_version.exists():
+                backup_file = self.file_to_version.with_suffix(
+                    f".json.{datetime.datetime.now().strftime('%Y%m%d%H%M%S')}-git-ignore-me.bak"
+                )
+                shutil.copyfile(self.file_to_version, backup_file)
 
             # copy the example file in place of the file_to_version
-            shutil.copyfile(self.example, self.file_to_version)
-            print(f"Updated {self.file_to_version} with new date.")
+            with self.file_to_version.open("w") as f:
+                json.dump(self.new_version, f, indent=2)
+                print(f"Updated {self.file_to_version} with new date.")
+        else:
+            print(f"No update needed for {self.file_to_version}.")
+
+        return needs_update
